memeprbarui Update dan Delete pada books, dengan bookId dan userId

+++ repositories/book_repository.go
@@ -9,11 +9,10 @@ import (
 
 type BookRepositoryInterface interface {
 	NewBook(ctx context.Context, newBook models.NewBook, idToken int) error
-	GetBookByIdLogin(ctx context.Context, bookId int, idToken int) (models.Book, error)
 	GetBookById(ctx context.Context, bookId int) (models.Book, error)
 	GetBookByTitle(ctx context.Context, title string) (models.Book, error)
 	GetAllBook(ctx context.Context) ([]models.Book, error)
-	DeleteBook(ctx context.Context, idToken int) error
+	DeleteBook(ctx context.Context, bookId int, idToken int) error
 	UpdateBook(ctx context.Context, updateBook models.Book, bookId int, idToken int) (models.Book, error)
 }
 
@@ -28,7 +27,7 @@ func NewBookRepository(db *sql.DB) *BookRepository {
 }
 
 func (br *BookRepository) NewBook(ctx context.Context, newBook models.NewBook, idToken int) error {
-	query := "INSERT INTO books (title, author, addedByUser) VALUES (?, ?, ?)"
+	query := "INSERT INTO books (title, author, userId) VALUES (?, ?, ?)"
 
 	_, err := br.mysql.ExecContext(ctx, query, newBook.Title, newBook.Author, idToken)
 	if err != nil {
@@ -38,25 +37,11 @@ func (br *BookRepository) NewBook(ctx context.Context, newBook models.NewBook, i
 	return nil
 }
 
-func (br *BookRepository) GetBookByIdLogin(ctx context.Context, bookId int, idToken int) (models.Book, error) {
-	var book models.Book
-	query := "SELECT bookId, title, author FROM books WHERE bookId = ?"
-
-	err := br.mysql.QueryRowContext(ctx, query, bookId).Scan(&book.BookId, &book.Title, &book.Author)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return models.Book{}, err
-		}
-		return models.Book{}, err
-	}
-	return book, nil
-}
-
 func (br *BookRepository) GetBookById(ctx context.Context, bookId int) (models.Book, error) {
 	var book models.Book
-	query := "SELECT bookId, title, author FROM books WHERE bookId = ?"
+	query := "SELECT bookId, title, author, userId FROM books WHERE bookId = ?"
 
-	err := br.mysql.QueryRowContext(ctx, query, bookId).Scan(&book.BookId, &book.Title, &book.Author)
+	err := br.mysql.QueryRowContext(ctx, query, bookId).Scan(&book.BookId, &book.Title, &book.Author, &book.UserId)
 	if err != nil {
 		if err == sql.ErrNoRows {
 			return models.Book{}, err
@@ -81,7 +66,7 @@ func (br *BookRepository) GetBookByTitle(ctx context.Context, title string) (mod
 }
 
 func (br *BookRepository) GetAllBook(ctx context.Context) ([]models.Book, error) {
-	query := "SELECT bookId, title, author, addedByUser FROM books"
+	query := "SELECT bookId, title, author, userId FROM books"
 
 	rows, err := br.mysql.QueryContext(ctx, query)
 	if err != nil {
@@ -92,7 +77,7 @@ func (br *BookRepository) GetAllBook(ctx context.Context) ([]models.Book, error)
 	var books []models.Book
 	for rows.Next() {
 		var book models.Book
-		err := rows.Scan(&book.BookId, &book.Title, &book.Author, &book.AddedByUser)
+		err := rows.Scan(&book.BookId, &book.Title, &book.Author, &book.UserId)
 		if err != nil {
 			return nil, err
 		}
@@ -101,10 +86,10 @@ func (br *BookRepository) GetAllBook(ctx context.Context) ([]models.Book, error)
 	return books, nil
 }
 
-func (br *BookRepository) DeleteBook(ctx context.Context, idToken int) error {
-	query := "DELETE FROM books WHERE bookId = ?"
+func (br *BookRepository) DeleteBook(ctx context.Context, bookId int, idToken int) error {
+	query := "DELETE FROM books WHERE bookId = ? AND userId = ?"
 
-	result, err := br.mysql.ExecContext(ctx, query, idToken)
+	result, err := br.mysql.ExecContext(ctx, query, bookId, idToken)
 	if err != nil {
 		return err
 	}
@@ -117,9 +102,9 @@ func (br *BookRepository) DeleteBook(ctx context.Context, idToken int) error {
 }
 
 func (br *BookRepository) UpdateBook(ctx context.Context, updateBook models.Book, bookId int, idToken int) (models.Book, error) {
-	query := "UPDATE books SET title = ?, author = ?, updatedByUser = ? WHERE bookId = ?"
+	query := "UPDATE books SET title = ?, author = ? WHERE bookId = ? AND userId = ?"
 
-	result, err := br.mysql.ExecContext(ctx, query, updateBook.Title, updateBook.Author, idToken, bookId)
+	result, err := br.mysql.ExecContext(ctx, query, updateBook.Title, updateBook.Author, bookId, idToken)
 	if err != nil {
 		return models.Book{}, err
 	}
